Feat: started with the profile page

+++ hiram/lib/features/user_profile/presentation/userprofile_header.dart
@@ -33,7 +33,7 @@ class _UserProfileState extends State<UserProfile> {
           _userName = userData['name'] ?? 'Unknown User';
           _itemsRentedOut = userData['itemsRentedOut'] ?? 0;
           _itemsRenting = userData['itemsRenting'] ?? 0;
-          _credibilityScore = userData['credibilityScore'] ?? 0.0;
+          _credibilityScore = (userData['credibilityScore'] ?? 0).toDouble();
         });
       }
     }
